- working on ios now! - renaming to viewSelectors - changing imports
